+++ packages/core/src/helpers/chain.ts
@@ -28,10 +28,16 @@ export function getExplorerAddressLink(address: string, chainId: ChainId) {
       return `https://explorer-mainnet.maticvigil.com/address/${address}/transactions`
     case ChainId.Mumbai:
       return `https://explorer-mumbai.maticvigil.com/address/${address}/transactions`
+    case ChainId.Theta:
+      return `https://explorer.thetatoken.org/address/${address}`
+    case ChainId.ThetaTestnet:
+      return `https://testnet-explorer.thetatoken.org/address/${address}`
     case ChainId.Harmony:
       return `https://explorer.harmony.one/address/${address}`
     case ChainId.Moonriver:
       return `https://blockscout.moonriver.moonbeam.network/address/${address}/transactions`
+    case ChainId.Palm:
+      return `https://explorer.palm.io/address/${address}`
     case ChainId.Fantom:
       return `https://ftmscan.com/address/${address}`
   }
@@ -53,10 +59,16 @@ export function getExplorerTransactionLink(transactionHash: string, chainId: Cha
       return `https://explorer-mainnet.maticvigil.com/tx/${transactionHash}/internal-transactions`
     case ChainId.Mumbai:
       return `https://explorer-mumbai.maticvigil.com/tx/${transactionHash}/internal-transactions`
+    case ChainId.Theta:
+      return `https://explorer.thetatoken.org/tx/${transactionHash}`
+    case ChainId.ThetaTestnet:
+      return `https://testnet-explorer.thetatoken.org/tx/${transactionHash}`
     case ChainId.Harmony:
       return `https://explorer.harmony.one/tx/${transactionHash}`
     case ChainId.Moonriver:
       return `https://blockscout.moonriver.moonbeam.network/tx/${transactionHash}/internal-transactions`
+    case ChainId.Palm:
+      return `https://explorer.palm.io/tx/${transactionHash}`
     case ChainId.Fantom:
       return `https://ftmscan.com/tx/${transactionHash}`
   }