+++ packages/core/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @usedapp/core
 
+## 0.5.2
+
+### Patch Changes
+
+- cf8e28d: Add support for Theta
+
+## 0.5.1
+
+### Patch Changes
+
+- aa95ff4: Add Palm support
+
+## 0.5.0
+
+### Minor Changes
+
+- f5a5c47: Added local storage path to config
+
+## 0.4.8
+
+### Patch Changes
+
+- a94fae5: Add missing lodash-merge dependency
+
 ## 0.4.7
 
 ### Patch Changes